move stack component into separate folder
